LayoutTests/http/wpt/web-animations/interfaces/Animation/effect-expected.txt


FAIL effect is set correctly. Can't find variable: KeyframeEffectReadOnly
FAIL Clearing and setting Animation.effect should update the computed style of the target element assert_equals: animation is initially having an effect expected "100px" but got "auto"
